requirements.txt
pillow
python-telegram-bot
aiohttp==3.9.0b0
aiodns
Brotli